# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LuaBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scripts/{id}/whitelist:
    get:
      operationId: listWhitelist
      tags: [whitelist]
      summary: List whitelist entries for a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of whitelist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhitelistEntry"
    post:
      operationId: addToWhitelist
      tags: [whitelist]
      summary: Add a Discord user to script whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhitelistInput"
      responses:
        "201":
          description: Added whitelist entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
  /scripts/{id}/whitelist/{entryId}:
    delete:
      operationId: removeFromWhitelist
      tags: [whitelist]
      summary: Remove a Discord user from script whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    WhitelistEntry:
      type: object
      required: [id, scriptId, discordUserId, createdAt]
      properties:
        id:
          type: integer
        scriptId:
          type: integer
        discordUserId:
          type: string
        addedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    WhitelistInput:
      type: object
      required: [discordUserId]
      properties:
        discordUserId:
          type: string
        addedBy:
          type: string