# lib/api-zod/src/generated/types/whitelistInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * LuaBox API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhitelistInput {
  discordUserId: string;
  addedBy?: string;
}